Format and style items within content1 div
Made 1 div for the content1 section and divided it up into
two separate sections. The left includes a h1, p, and button
and the right includes a div that serves a placeholder for an
img. The size of each of the divs still needs to be increased a bit
and the margin/padding also needs to be decreased on the left side.