Finish formatting in content 1 and add content 2
Utilized chrome/edge dev tools to see where I was going wrong with the
size of the divs for the header and content1. Centered the header and
modified the sizes of each div to match them up with the content 1 divs.
I split content 2 up into multiple nested divs. I used divs with borders
to make the empty image boxes and nested the text in with those same divs.
From there, I was able to center the divs easily and then add padding
around the image/text divs to give them space between each other and to
add padding between the boxes and the upper and lower parts.