Finish formatting and styling content3 section
I split this section up into the main section div and then a smaller div
that holds the quote (h1) and the author (h2). From there, everything
was centered and the author part was aligned to the rightmost part of
its parent div. A bit of padding was added to make it more closely match
the example.